# docs/openapi.json
{
  "components": {
    "schemas": {
      "AccountDetail": {
        "properties": {
          "account_id": {
            "title": "Account Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "device_id": {
            "title": "Device Id",
            "type": "string"
          },
          "email_hash": {
            "title": "Email Hash",
            "type": "string"
          },
          "features": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Features",
            "type": "object"
          },
          "ip_address": {
            "title": "Ip Address",
            "type": "string"
          },
          "ml_score": {
            "title": "Ml Score",
            "type": "number"
          },
          "payment_instrument_id": {
            "title": "Payment Instrument Id",
            "type": "string"
          },
          "phone_hash": {
            "title": "Phone Hash",
            "type": "string"
          },
          "predicted_label": {
            "title": "Predicted Label",
            "type": "integer"
          },
          "reason_codes": {
            "items": {
              "type": "string"
            },
            "title": "Reason Codes",
            "type": "array"
          },
          "ring_ids": {
            "items": {
              "type": "string"
            },
            "title": "Ring Ids",
            "type": "array"
          },
          "transactions": {
            "items": {
              "$ref": "#/components/schemas/TransactionOut"
            },
            "title": "Transactions",
            "type": "array"
          }
        },
        "required": [
          "account_id",
          "created_at",
          "email_hash",
          "phone_hash",
          "device_id",
          "ip_address",
          "payment_instrument_id",
          "ml_score",
          "predicted_label",
          "reason_codes",
          "features",
          "ring_ids",
          "transactions"
        ],
        "title": "AccountDetail",
        "type": "object"
      },
      "AnalyzeResponse": {
        "properties": {
          "account_count": {
            "title": "Account Count",
            "type": "integer"
          },
          "labels_available": {
            "title": "Labels Available",
            "type": "boolean"
          },
          "ring_count": {
            "title": "Ring Count",
            "type": "integer"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "transaction_count": {
            "title": "Transaction Count",
            "type": "integer"
          }
        },
        "required": [
          "run_id",
          "account_count",
          "transaction_count",
          "ring_count",
          "labels_available"
        ],
        "title": "AnalyzeResponse",
        "type": "object"
      },
      "Body_analyze_api_v1_analyze_post": {
        "properties": {
          "accounts": {
            "contentMediaType": "application/octet-stream",
            "description": "accounts.csv, UTF-8",
            "title": "Accounts",
            "type": "string"
          },
          "transactions": {
            "contentMediaType": "application/octet-stream",
            "description": "transactions.csv, UTF-8",
            "title": "Transactions",
            "type": "string"
          }
        },
        "required": [
          "accounts",
          "transactions"
        ],
        "title": "Body_analyze_api_v1_analyze_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReviewStatusUpdate": {
        "properties": {
          "status": {
            "enum": [
              "new",
              "reviewing",
              "confirmed",
              "dismissed"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "ReviewStatusUpdate",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Razorpay Abuse Ring Detector",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/accounts/{account_id}": {
      "get": {
        "operationId": "account_detail_api_v1_accounts__account_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "title": "Account Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Account Detail",
        "tags": [
          "review"
        ]
      }
    },
    "/api/v1/analyze": {
      "post": {
        "operationId": "analyze_api_v1_analyze_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_analyze_api_v1_analyze_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyzeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Analyze",
        "tags": [
          "review"
        ]
      }
    },
...
    "/api/v1/rings": {
      "get": {
        "operationId": "list_rings_api_v1_rings_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "description": "Minimum ring score, inclusive.",
            "in": "query",
            "name": "min_score",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Minimum ring score, inclusive.",
              "maximum": 1,
              "minimum": 0,
              "title": "Min Score",
              "type": "number"
            }
          },
          {
            "description": "Exact review status.",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "new",
                    "reviewing",
                    "confirmed",
                    "dismissed"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Exact review status.",
              "title": "Status"
            }
          },
          {
            "description": "Exact promotion ID.",
            "in": "query",
            "name": "promotion",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Exact promotion ID.",
              "title": "Promotion"
            }
          },
          {
            "description": "Earliest member signup date, inclusive.",
            "in": "query",
            "name": "date_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Earliest member signup date, inclusive.",
              "title": "Date From"
            }
          },
          {
            "description": "Latest member signup date, inclusive.",
            "in": "query",
            "name": "date_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Latest member signup date, inclusive.",
              "title": "Date To"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RingPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Rings",
        "tags": [
          "review"
        ]
      }
    },
    "/api/v1/rings/{ring_id}": {
      "get": {
        "operationId": "ring_detail_api_v1_rings__ring_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "ring_id",
            "required": true,
            "schema": {
              "title": "Ring Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RingDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ring Detail",
        "tags": [
          "review"
        ]
      }
    },
    "/api/v1/rings/{ring_id}/status": {
      "patch": {
        "operationId": "update_ring_status_api_v1_rings__ring_id__status_patch",
        "parameters": [
          {
            "in": "path",
            "name": "ring_id",
            "required": true,
            "schema": {
              "title": "Ring Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewStatusUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Ring Status",
        "tags": [
          "review"
        ]
      }
    },
...
  }
}